refactor(PatternRow, FormattedDuration & Time): Localize PatternRow, move BoldText, FormattedDuratio
