+++ i18n/en-US.yml
@@ -152,6 +152,9 @@ components:
     emailSelect: Email
     noneSelect: Don't notify me
     smsSelect: SMS
+  PatternRow:
+    routeName: "<b>{routeName}</b> To {headsign}"
+    routeShort: "To {headsign}"
   PhoneNumberEditor:
     changeNumber: Change number
     invalidCode: Please enter 6 digits for the validation code.
@@ -416,6 +419,8 @@ common:
   # Shared itinerary description messages
   itineraryDescriptions:
     calories: Cal
+    departure: DEPARTURE
+    status: STATUS
     noItineraryToDisplay: No itinerary to display.
     transfers: "{transfers, plural, =0 {} one {{transfers} transfer} other {{transfers} transfers}}"
