+++ lib/components/narrative/default/default-itinerary.js
@@ -4,13 +4,13 @@ import { FormattedMessage, FormattedNumber } from 'react-intl'
 import { connect } from 'react-redux'
 
 import FieldTripGroupSize from '../../admin/field-trip-itinerary-group-size'
+import { FormattedDuration } from '../../util/format-duration'
+import { FormattedTime } from '../../util/format-time'
 import NarrativeItinerary from '../narrative-itinerary'
 import ItineraryBody from '../line-itin/connected-itinerary-body'
 import SimpleRealtimeAnnotation from '../simple-realtime-annotation'
 
 import ItinerarySummary from './itinerary-summary'
-import { FormattedDuration } from './format-duration'
-import { FormattedTime } from './format-time'
 
 const { calculateFares, isBicycle, isMicromobility, isTransit } = coreUtils.itinerary
