+++ lib/components/narrative/tabbed-itineraries.js
@@ -7,10 +7,9 @@ import { connect } from 'react-redux'
 
 import * as narrativeActions from '../../actions/narrative'
 import { ComponentContext } from '../../util/contexts'
+import { FormattedDuration } from '../util/format-duration'
+import { FormattedTime } from '../util/format-time'
 import { getActiveSearch, getRealtimeEffects } from '../../util/state'
-
-import { FormattedDuration } from './default/format-duration'
-import { FormattedTime } from './default/format-time'
 
 const { calculateFares, calculatePhysicalActivity } = coreUtils.itinerary
