+++ lib/components/user/monitored-trip/trip-status-rendering-strategies/active-trip-renderer.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { FormattedMessage } from 'react-intl'
 
 import { getTripStatus, REALTIME_STATUS } from '../../../../util/viewer'
-import { FormattedDuration } from '../../../narrative/default/format-duration'
+import { FormattedDuration } from '../../../util/format-duration'
 
 import baseRenderer from './base-renderer'
