+++ lib/components/user/monitored-trip/trip-status-rendering-strategies/base-renderer.js
@@ -1,7 +1,7 @@
 import moment from 'moment'
 import { FormattedMessage } from 'react-intl'
 
-import { FormattedDuration } from '../../../narrative/default/format-duration'
+import { FormattedDuration } from '../../../util/format-duration'
 
 /**
  * Calculate commonly-used pieces of data used to render the trip status
