+++ lib/components/user/monitored-trip/trip-status-rendering-strategies/upcoming-trip-renderer.js
@@ -2,7 +2,7 @@ import moment from 'moment'
 import React from 'react'
 import { FormattedMessage } from 'react-intl'
 
-import { FormattedDuration } from '../../../narrative/default/format-duration'
+import { FormattedDuration } from '../../../util/format-duration'
 import { getMinutesUntilItineraryStart } from '../../../../util/itinerary'
 import { getTripStatus, REALTIME_STATUS } from '../../../../util/viewer'
