+++ lib/components/user/monitored-trip/trip-summary-pane.js
@@ -2,10 +2,9 @@ import React from 'react'
 import { FormattedList, FormattedMessage } from 'react-intl'
 
 import { dayFieldsToArray } from '../../../util/monitored-trip'
+import { BoldText } from '../../../util/i18n'
 
 import TripSummary from './trip-summary'
-
-const BoldText = contents => <b>{contents}</b>
 
 /**
  * Displays the summary information of a monitored trip.
