+++ lib/components/user/monitored-trip/trip-summary.js
@@ -3,8 +3,8 @@ import React from 'react'
 import { FormattedMessage } from 'react-intl'
 import { connect } from 'react-redux'
 
-import { FormattedDuration } from '../../narrative/default/format-duration'
-import { FormattedTime } from '../../narrative/default/format-time'
+import { FormattedDuration } from '../../util/format-duration'
+import { FormattedTime } from '../../util/format-time'
 import ItinerarySummary from '../../narrative/default/itinerary-summary'
 
 const TripSummary = ({ monitoredTrip, use24HourFormat }) => {
