+++ lib/components/util/format-duration.js
