+++ lib/components/viewers/pattern-row.js
@@ -1,6 +1,8 @@
 import React, { Component } from 'react'
+import {FormattedMessage} from 'react-intl'
 import { VelocityTransitionGroup } from 'velocity-react'
 
+import { BoldText } from '../../util/i18n'
 import Icon from '../narrative/icon'
 import { stopTimeComparator } from '../../util/viewer'
 
@@ -56,7 +58,14 @@ export default class PatternRow extends Component {
         <div className='header'>
           {/* route name */}
           <div className='route-name'>
-            <b>{routeName}</b> To {pattern.headsign}
+            <FormattedMessage
+              id='components.PatternRow.routeName'
+              values={{
+                b: BoldText,
+                headsign: pattern.headsign,
+                routeName: routeName
+              }}
+            />
           </div>
           {/* next departure preview */}
           {hasStopTimes && (
@@ -88,8 +97,12 @@ export default class PatternRow extends Component {
                 {/* trips table header row */}
                 <div className='header'>
                   <div className='cell' />
-                  <div className='cell time-column'>DEPARTURE</div>
-                  <div className='cell status-column'>STATUS</div>
+                  <div className='cell time-column'><FormattedMessage
+                    id='common.itineraryDescriptions.departure'
+                  /> </div>
+                  <div className='cell status-column'><FormattedMessage
+                    id='common.itineraryDescriptions.status'
+                  /></div>
                 </div>
 
                 {/* list of upcoming trips */}
@@ -101,9 +114,10 @@ export default class PatternRow extends Component {
                         className='trip-row'
                         key={i}
                         style={{ display: 'table-row', fontSize: 14, marginTop: 6 }}>
-                        <div className='cell'>
-                          To {stopTime.headsign}
-                        </div>
+                        <div className='cell'> <FormattedMessage
+                          id='components.PatternRow.routeShort'
+                          values={{headsign: stopTime.headsign}}
+                        /></div>
                         <div className='cell time-column'>
                           <StopTimeCell
                             homeTimezone={homeTimezone}
