+++ lib/components/viewers/stop-viewer.js
@@ -12,14 +12,12 @@ import styled from 'styled-components'
 import * as apiActions from '../../actions/api'
 import * as mapActions from '../../actions/map'
 import * as uiActions from '../../actions/ui'
+import { BoldText, StrongText } from '../../util/i18n'
 import Icon from '../narrative/icon'
 import { getShowUserSettings, getStopViewerConfig } from '../../util/state'
 
 import LiveStopTimes from './live-stop-times'
 import StopScheduleTable from './stop-schedule-table'
-
-const BoldText = contents => <b>{contents}</b>
-const StrongText = contents => <strong>{contents}</strong>
 
 const {
   getUserTimezone,
