+++ lib/util/i18n.js
@@ -64,3 +64,13 @@ export function getDefaultLocale (config) {
   const { localization = {} } = config
   return localization.defaultLocale || 'en-US'
 }
+
+/**
+ * BoldText function component for injecting <b> formatting into i18n messages.
+ */
+export const BoldText = contents => <b>{contents}</b>
+
+/**
+ * StrongText function component for injecting <strong> formatting into i18n messages.
+ */
+export const StrongText = contents => <strong>{contents}</strong>
